refactor: Http11Method 열거형 구현
