+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -7,10 +7,10 @@ import java.io.InputStreamReader;
 import java.net.Socket;
 import java.util.LinkedList;
 import java.util.List;
+import org.apache.Serializer;
 import org.apache.coyote.HttpResponse;
 import org.apache.coyote.Processor;
 import org.apache.coyote.RequestHandler;
-import org.apache.coyote.Serializer;
 import org.apache.coyote.handler.LoginRequestHandler;
 import org.apache.coyote.handler.NotFoundHandler;
 import org.apache.coyote.handler.ResourceRequestHandler;
